[Decomp][Spell] Cohesion split: extract SpellTargetList handlers to SpellTargetList.cpp
